adicionando tratamento de erros e feedback ao src
