+++ .gitignore
@@ -1,3 +1,5 @@
 .vscode
 target
 *.pdf
+**/__pycache__/
+front
