+++ src/main.py
@@ -1,5 +1,5 @@
 from compiler.compiler import Compiler
-from syntatic.semantyc import AnalisadorSemantico
+from semantic.semantyc import AnalisadorSemantico
 from syntatic.TreeGenerator import NoInterno, NoFolha
 
 import sys
