+++ src/read.me
@@ -1,0 +1,35 @@
+# Executando o Programa
+
+Este documento fornece instruções sobre como executar o programa. Siga as etapas abaixo para obter os resultados desejados.
+
+## Pré-requisitos
+
+Certifique-se de ter Python 3 instalado em seu sistema. Você pode verificar a versão do Python em seu sistema executando o seguinte comando no terminal:
+
+\```bash
+python3 --version
+\```
+
+## Como Executar
+
+1. Navegue até o diretório onde o arquivo `main.py` está localizado.
+
+2. Execute o seguinte comando no terminal:
+
+3. lembre-se de estar dentro da pasta src no terminal.
+
+\```bash
+python3 main.py semantic/test/error4semantic.w
+\```
+
+Neste comando, `semantic/test/error4semantic.w` é o caminho para o arquivo de teste que você deseja usar. Certifique-se de substituir `semantic/test/error4semantic.w` pelo caminho do arquivo de teste que você deseja usar.
+
+## Resultados Esperados
+
+Após a execução bem-sucedida do programa, uma tabela de tokens será impressa no console. 
+
+## Tratamento de Erros
+
+Caso haja algum erro durante a execução do programa, os analisadores irão identificar e referenciar o erro. 
+
+Por favor, revise seu arquivo de teste e tente novamente.
