+++ src/semantic/test/error1syntatic.w
@@ -1,0 +1,9 @@
+programa "exemplo3":
+inicio
+      x = (3 ^ 5) ^ 2
+      y = 3 ^ 5 ^ 2
+      z = (10 + 2 * 50 / 5 - 1) % 2
+      se (3 + 4 - 5) % 2 = 0 entao inicio
+      mostrar(10 / 2)
+      fim
+fim.
