+++ src/semantic/test/error2syntatic.w
@@ -1,0 +1,9 @@
+programa "exemplo4":
+inicio
+      x = (3 ^ 5) ^ 2
+      y = 3 ^ 5 ^ 2
+      z = (10 + 2 * 50 / 5 - 1) % 2
+      se (3 + 4 -5) % 2 == 0 entao inicio
+      mostrar(10 / 2, 2 * 1)
+      fim
+fim.
