+++ src/semantic/test/error3semantic.w
@@ -1,0 +1,4 @@
+programa "teste semântico 1":
+inicio
+      x = y
+fim.
