+++ src/semantic/test/error4semantic.w
@@ -1,0 +1,7 @@
+programa "teste semântico 2":
+inicio
+      x = 1
+      y = 2
+      mostrar(z)
+      z = y
+fim.
